requirements.txt
flask
gunicorn
google-cloud-secret-manager